code refractoring
